+++ kubernetes-tests/src/test/java/io/fabric8/kubernetes/client/mock/NamespaceTest.java
@@ -25,14 +25,17 @@ import io.fabric8.kubernetes.client.KubernetesClientException;
 import io.fabric8.kubernetes.client.server.mock.KubernetesServer;
 import io.fabric8.kubernetes.client.utils.Utils;
 import org.junit.Rule;
-import org.junit.Test;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.migrationsupport.rules.EnableRuleMigrationSupport;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNull;
 
+@EnableRuleMigrationSupport
 public class NamespaceTest {
   @Rule
   public KubernetesServer server = new KubernetesServer();
@@ -74,12 +77,14 @@ public class NamespaceTest {
     assertEquals(0, namespaceList.getItems().size());
   }
 
-  @Test(expected = KubernetesClientException.class)
+  @Test
   public void testEditMissing() {
-    server.expect().withPath("/api/v1/namespaces/namespace1").andReturn(404, "error message from kubernetes").always();
-    KubernetesClient client = server.getClient();
+    Assertions.assertThrows(KubernetesClientException.class, () -> {
+      server.expect().withPath("/api/v1/namespaces/namespace1").andReturn(404, "error message from kubernetes").always();
+      KubernetesClient client = server.getClient();
 
-    client.namespaces().withName("namespace1").edit();
+      client.namespaces().withName("namespace1").edit();
+    });
   }
 
   @Test