+++ kubernetes-tests/src/test/java/io/fabric8/openshift/client/server/mock/KubernetesOperationTest.java
@@ -24,10 +24,12 @@ import io.fabric8.openshift.api.model.BuildConfigBuilder;
 import io.fabric8.openshift.client.OpenShiftClient;
 
 import org.junit.Rule;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.migrationsupport.rules.EnableRuleMigrationSupport;
 
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
+@EnableRuleMigrationSupport
 public class KubernetesOperationTest {
   @Rule
   public OpenShiftServer server = new OpenShiftServer();